Added styles to listing page
